Connor's Data Parser
I added an icon for Janna lol.

I re-created the logic you made in the "transcription" folder in a folder I called "connor-data-handler" There's a button to upload a file, or you can push "use test data" which will load the conversation .json you had in the project folder.

I have two different parsing steps, first parsing the facebook data. I collate continuous messages from the same person into one message. I also added the feature where the identified "I" nouns are replaced with the name of the sender.

Then I have a function for creating edges based on a list of groups of nodes. Essentially I turn the message exchange (a list of messages) into a list of groups of nouns. I then create an edge that links each noun to the nouns it appears with, and then edges between nouns that appear in sequential messages.

To uniqueify the edges I created an ID for each edge based on what two nodes it's relating, then I weight the nodes based on how many times that node appears in the set.

The next step for me is to make a reasonably nice looking way to display all the nodes, then when you click on a node, or hover or whatever you get some nice display of the things it's connected to.

Then we can look into what the best way to save this data is going to be.